dht11 - Publisher/Provider-Template-Pi4All2Localhost/requirements.txt
Pygments>=2.9.0
paho-mqtt>=1.5.1
requests>=2.25.1
termcolor>=1.1.0
urllib3>=1.26.6
sense-hat>=1.16.2